Consolidate Transaction Atom and Payload Atom into a single Atom
