+++ radixdlt-java/src/main/java/com/radixdlt/client/application/translate/DataStoreTranslator.java
@@ -2,8 +2,8 @@ package com.radixdlt.client.application.translate;
 
 import com.radixdlt.client.application.actions.DataStore;
 import com.radixdlt.client.application.objects.Data;
-import com.radixdlt.client.core.atoms.ApplicationPayloadAtom;
 import com.radixdlt.client.core.atoms.AtomBuilder;
+import com.radixdlt.client.core.atoms.TransactionAtom;
 import com.radixdlt.client.core.crypto.EncryptedPrivateKey;
 import io.reactivex.Completable;
 import java.util.Collections;
@@ -22,7 +22,7 @@ public class DataStoreTranslator {
 	}
 
 	public Completable translate(DataStore dataStore, AtomBuilder atomBuilder) {
-		atomBuilder.type(ApplicationPayloadAtom.class);
+		atomBuilder.type(TransactionAtom.class);
 		atomBuilder.payload(dataStore.getData().getBytes());
 
 		if (!dataStore.getData().getProtectors().isEmpty()) {
@@ -38,7 +38,7 @@ public class DataStoreTranslator {
 		return Completable.complete();
 	}
 
-	public Data fromAtom(ApplicationPayloadAtom atom) {
+	public Data fromAtom(TransactionAtom atom) {
 		final List<EncryptedPrivateKey> protectors;
 		if (atom.getEncryptor() != null && atom.getEncryptor().getProtectors() != null) {
 			protectors = atom.getEncryptor().getProtectors();
