+++ radixdlt-java/src/main/java/com/radixdlt/client/core/atoms/AtomBuilder.java
@@ -110,10 +110,7 @@ public class AtomBuilder {
 
 		final Atom atom;
 		if (TransactionAtom.class.isAssignableFrom(atomClass)) {
-			atom = new TransactionAtom(particles, destinations, payload, encryptor, this.timestamp);
-		} else if (ApplicationPayloadAtom.class.isAssignableFrom(atomClass)) {
-			Objects.requireNonNull(applicationId, "Payload Atom must have an application id");
-			atom = new ApplicationPayloadAtom(applicationId, particles, destinations, payload, encryptor, this.timestamp);
+			atom = new TransactionAtom(applicationId, particles, destinations, payload, encryptor, this.timestamp);
 		} else {
 			throw new IllegalStateException("Unable to create atom with class: " + atomClass.getSimpleName());
 		}
