+++ radixdlt-java/src/test/java/com/radixdlt/client/core/network/RadixJsonRpcClientTest.java
@@ -11,9 +11,9 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.radixdlt.client.core.address.EUID;
-import com.radixdlt.client.core.atoms.ApplicationPayloadAtom;
 import com.radixdlt.client.core.atoms.Atom;
 import com.radixdlt.client.core.atoms.Shards;
+import com.radixdlt.client.core.atoms.TransactionAtom;
 import com.radixdlt.client.core.network.AtomSubmissionUpdate.AtomSubmissionState;
 import com.radixdlt.client.core.network.WebSocketClient.RadixClientStatus;
 import com.radixdlt.client.core.serialization.RadixJson;
@@ -142,7 +142,7 @@ public class RadixJsonRpcClientTest {
 			String id = jsonObject.get("id").getAsString();
 
 			JsonArray atoms = new JsonArray();
-			Atom atom = new ApplicationPayloadAtom("Test", null, null, null, null, 1);
+			Atom atom = new TransactionAtom("Test", null, null, null, null, 1);
 			atoms.add(gson.toJsonTree(atom, Atom.class));
 
 			JsonObject response = new JsonObject();
@@ -158,7 +158,7 @@ public class RadixJsonRpcClientTest {
 
 		jsonRpcClient.getAtom(new EUID(1)).subscribe(observer);
 
-		observer.assertValue(atom -> atom.getAsMessageAtom().getApplicationId().equals("Test"));
+		observer.assertValue(atom -> atom.getAsTransactionAtom().getApplicationId().equals("Test"));
 		observer.assertComplete();
 		observer.assertNoErrors();
 	}
@@ -194,7 +194,7 @@ public class RadixJsonRpcClientTest {
 
 			JsonArray atoms = new JsonArray();
 			JsonElement atom = gson.toJsonTree(
-				new ApplicationPayloadAtom("Test", null, null, null, null, 1),
+				new TransactionAtom("Test", null, null, null, null, 1),
 				Atom.class
 			);
 			atoms.add(atom);
@@ -207,13 +207,13 @@ public class RadixJsonRpcClientTest {
 		}).when(wsClient).send(any());
 		RadixJsonRpcClient jsonRpcClient = new RadixJsonRpcClient(wsClient);
 
-		TestObserver<ApplicationPayloadAtom> observer = new TestObserver<>();
+		TestObserver<TransactionAtom> observer = new TestObserver<>();
 
-		jsonRpcClient.getAtoms(new AtomQuery<>(new EUID(1), ApplicationPayloadAtom.class)).subscribe(observer);
+		jsonRpcClient.getAtoms(new AtomQuery<>(new EUID(1), TransactionAtom.class)).subscribe(observer);
 
 		observer.assertNoErrors();
 		observer.assertValueCount(1);
-		observer.assertValue(atom -> atom.getAsMessageAtom().getApplicationId().equals("Test"));
+		observer.assertValue(atom -> atom.getAsTransactionAtom().getApplicationId().equals("Test"));
 	}
 
 	@Test
@@ -249,7 +249,7 @@ public class RadixJsonRpcClientTest {
 				params.addProperty("subscriberId", subscriberId);
 				JsonArray atoms = new JsonArray();
 				JsonElement atom = gson.toJsonTree(
-					new ApplicationPayloadAtom("Test", null, null, null, null, 1),
+					new TransactionAtom("Test", null, null, null, null, 1),
 					Atom.class
 				);
 				atoms.add(atom);
@@ -264,9 +264,9 @@ public class RadixJsonRpcClientTest {
 		}).when(wsClient).send(any());
 		RadixJsonRpcClient jsonRpcClient = new RadixJsonRpcClient(wsClient);
 
-		TestObserver<ApplicationPayloadAtom> observer = new TestObserver<>();
+		TestObserver<TransactionAtom> observer = new TestObserver<>();
 
-		jsonRpcClient.getAtoms(new AtomQuery<>(new EUID(1), ApplicationPayloadAtom.class))
+		jsonRpcClient.getAtoms(new AtomQuery<>(new EUID(1), TransactionAtom.class))
 			.subscribe(observer);
 		observer.cancel();
 
@@ -319,7 +319,7 @@ public class RadixJsonRpcClientTest {
 		TestObserver<AtomSubmissionUpdate> observer = new TestObserver<>();
 
 		jsonRpcClient.submitAtom(
-			new ApplicationPayloadAtom("Test", null, null, null, null, 1)
+			new TransactionAtom("Test", null, null, null, null, 1)
 		).subscribe(observer);
 
 		observer.assertNoErrors();
